+++ plugins/module_utils/oci_custom_helpers_utils.py
@@ -82,6 +82,7 @@ from ansible_collections.oracle.oci.plugins.module_utils import (
     oci_network_load_balancer_custom_helpers,
     oci_service_catalog_custom_helpers,
     oci_golden_gate_custom_helpers,
+    oci_generic_artifacts_content_custom_helpers,
 )  # noqa
 
 custom_helper_mapping = get_custom_class_mapping(
@@ -115,6 +116,7 @@ custom_helper_mapping = get_custom_class_mapping(
         oci_waas_custom_helpers,
         oci_data_flow_custom_helpers,
         oci_analytics_custom_helpers,
+        oci_generic_artifacts_content_custom_helpers,
         oci_data_science_custom_helpers,
         oci_data_safe_custom_helpers,
         oci_data_catalog_custom_helpers,