+++ plugins/modules/oci_database_maintenance_run_facts.py
@@ -211,6 +211,18 @@ maintenance_runs:
             returned: on success
             type: string
             sample: "ocid1.peermaintenancerun.oc1..xxxxxxEXAMPLExxxxxx"
+        patching_mode:
+            description:
+                - "Maintenance method, it will be either \\"ROLLING\\" or \\"NONROLLING\\". Default value is ROLLING."
+            returned: on success
+            type: string
+            sample: ROLLING
+        patch_failure_count:
+            description:
+                - Contain the patch failure count.
+            returned: on success
+            type: int
+            sample: 56
     sample: [{
         "id": "ocid1.resource.oc1..xxxxxxEXAMPLExxxxxx",
         "compartment_id": "ocid1.compartment.oc1..xxxxxxEXAMPLExxxxxx",
@@ -226,7 +238,9 @@ maintenance_runs:
         "maintenance_type": "PLANNED",
         "patch_id": "ocid1.patch.oc1..xxxxxxEXAMPLExxxxxx",
         "maintenance_subtype": "QUARTERLY",
-        "peer_maintenance_run_id": "ocid1.peermaintenancerun.oc1..xxxxxxEXAMPLExxxxxx"
+        "peer_maintenance_run_id": "ocid1.peermaintenancerun.oc1..xxxxxxEXAMPLExxxxxx",
+        "patching_mode": "ROLLING",
+        "patch_failure_count": 56
     }]
 """