+++ plugins/modules/oci_limits_resource_availability_facts.py
@@ -24,9 +24,9 @@ short_description: Fetches details about a ResourceAvailability resource in Orac
 description:
     - Fetches details about a ResourceAvailability resource in Oracle Cloud Infrastructure
     - "For a given compartmentId, resource limit name, and scope, returns the following:
-        - the number of available resources associated with the given limit
-        - the usage in the selected compartment for the given limit
-        Note: not all resource limits support this API. If the value is not available, the API will return 404."
+        * The number of available resources associated with the given limit.
+        * The usage in the selected compartment for the given limit.
+        Note that not all resource limits support this API. If the value is not available, the API returns a 404 response."
 version_added: "2.9"
 author: Oracle (@oracle)
 options:
@@ -49,7 +49,7 @@ options:
         description:
             - "This field is mandatory if the scopeType of the target resource limit is AD.
               Otherwise, this field should be omitted.
-              If the above requirements are not met, the API will return a 400 - InvalidParameter response."
+              If the above requirements are not met, the API returns a 400 - InvalidParameter response."
         type: str
 extends_documentation_fragment: [ oracle.oci.oracle ]
 """
@@ -72,15 +72,15 @@ resource_availability:
     contains:
         used:
             description:
-                - The current usage in the given compartment. Because we have introduced resources with fractional counts,
-                  the field will round up to the nearest integer.
+                - The current usage in the given compartment. To support resources with fractional counts,
+                  the field rounds up to the nearest integer.
             returned: on success
             type: int
             sample: 56
         available:
             description:
-                - The count of available resources. Because we have introduced resources with fractional counts,
-                  the field will round down to the nearest integer.
+                - The count of available resources. To support resources with fractional counts,
+                  the field rounds down to the nearest integer.
             returned: on success
             type: int
             sample: 56
@@ -98,7 +98,7 @@ resource_availability:
             sample: 10
         effective_quota_value:
             description:
-                - The effective quota value for given compartment. This field is only present if there is a
+                - The effective quota value for the given compartment. This field is only present if there is a
                   current quota policy affecting the current resource in the target region or availability domain.
             returned: on success
             type: float